int to float

+++ packedForest/src/baseFunctions/fpForest.h
@@ -146,7 +146,7 @@ namespace fp {
 					return OOBaccuracy;
 				}
 
-                inline std::map<std::pair<int, int>, int> returnPairMat(){
+                inline std::map<std::pair<int, int>, double> returnPairMat(){
                                     return forest->returnPairMat();
                             }
 				float testAccuracy(){
