+++ packedForest/src/baseFunctions/fpForestBase.h
@@ -29,7 +29,7 @@ namespace fp{
 				virtual std::vector<int> predictClassPost(std::vector<T>& observation) = 0;
 				virtual int predictClass(const T* observation) = 0;
 				virtual float reportOOB() = 0; //TODO: JLP, finish this implementation.
-				virtual std::map<std::pair<int, int>, int> returnPairMat() = 0;
+				virtual std::map<std::pair<int, int>, double> returnPairMat() = 0;
 		};
 
 }//namespace fp
